Add identifier.h to capture middleware transporter identifer string creation and parsing
