+++ src/udpm/transport/intermodule.h
@@ -1,4 +1,4 @@
-// Copyright 2020-2021:
+// Copyright 2020-2026:
 //   GobySoft, LLC (2013-)
 //   Community contributors (see AUTHORS file)
 // File authors:
